Hacked compile error by casting to 32bit - doh!

+++ common/cmd_bootm.c
@@ -674,7 +674,7 @@ static int booti_setup(bootm_headers_t *images)
 
 		src = (void *)images->ep;
 		images->ep = dst;
-		memmove((void *)dst, src, le32_to_cpu(ih->image_size));
+		memmove((void *)(uint32_t)dst, src, le32_to_cpu(ih->image_size));
 	}
 
 	return 0;
